# contracts/openapi-v1.json
{
  "components": {
    "schemas": {
      "AddFindingRequest": {
        "properties": {
          "component": {
            "default": "",
            "maxLength": 200,
            "title": "Component",
            "type": "string"
          },
          "lead_time_weeks": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lead Time Weeks"
          },
          "parameter": {
            "default": "",
            "maxLength": 200,
            "title": "Parameter",
            "type": "string"
          },
          "predicted_cx_test": {
            "default": "",
            "maxLength": 200,
            "title": "Predicted Cx Test",
            "type": "string"
          },
          "provided_value": {
            "default": "",
            "maxLength": 500,
            "title": "Provided Value",
            "type": "string"
          },
          "rationale": {
            "default": "",
            "maxLength": 2000,
            "title": "Rationale",
            "type": "string"
          },
          "required_value": {
            "default": "",
            "maxLength": 500,
            "title": "Required Value",
            "type": "string"
          },
          "severity": {
            "default": "",
            "maxLength": 50,
            "title": "Severity",
            "type": "string"
          },
          "spec_clause": {
            "default": "",
            "maxLength": 200,
            "title": "Spec Clause",
            "type": "string"
          },
          "standard_ref": {
            "default": "",
            "maxLength": 200,
            "title": "Standard Ref",
            "type": "string"
          },
          "unit": {
            "default": "",
            "maxLength": 50,
            "title": "Unit",
            "type": "string"
          }
        },
        "title": "AddFindingRequest",
        "type": "object"
      },
...
      "CreateCaseRequest": {
        "properties": {
          "name": {
            "default": "",
            "maxLength": 200,
            "title": "Name",
            "type": "string"
          }
        },
        "title": "CreateCaseRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "UpdateFindingRequest": {
        "properties": {
          "owner": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner"
          },
          "resolution_note": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution Note"
          },
          "status": {
            "anyOf": [
              {
                "maxLength": 30,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          }
        },
        "title": "UpdateFindingRequest",
        "type": "object"
      },
      "UpdateRfiRequest": {
        "properties": {
          "response_text": {
            "anyOf": [
              {
                "maxLength": 4000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Response Text"
          },
          "status": {
            "maxLength": 30,
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "UpdateRfiRequest",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Spec-to-Site Deviation Sentinel for hyperscale data-centre EPC delivery",
    "title": "Pramaan — EPC Deviation Intelligence",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/cases": {
      "post": {
        "description": "Create a new case. The returned `secret` is shown exactly once — it\nis the only credential for every /cases/{case_id}/* route on this case\nand cannot be recovered if lost, the same way a demo token can't be.",
        "operationId": "v1_create_case",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCaseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "V1 Create Case",
        "tags": [
          "v1"
        ]
      }
    },
    "/api/v1/cases/{case_id}": {
      "delete": {
        "description": "Permanently delete a case and all its findings, RFIs, and audit log.\n\nRequires the X-Case-Secret header. Returns 404 for wrong/missing secret\n(indistinguishable from a nonexistent case).",
        "operationId": "v1_delete_case_endpoint",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "V1 Delete Case Endpoint",
        "tags": [
          "v1"
        ]
      },
      "get": {
        "operationId": "v1_get_case",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "V1 Get Case",
        "tags": [
          "v1"
        ]
      }
    },
    "/api/v1/cases/{case_id}/audit-log": {
      "get": {
        "operationId": "v1_get_case_audit_log",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "V1 Get Case Audit Log",
        "tags": [
          "v1"
        ]
      }
    },
    "/api/v1/cases/{case_id}/export/itp.pdf": {
      "get": {
        "description": "A physically-signable Inspection & Test Plan, distinct from the\nJSON/HTML exports — a commissioning authority can print and hand this\noff rather than work from a screen. Built from the case's own findings\nagainst the same commissioning-test catalog `backend/agents/commissioning.py`\nuses; adds no new AI reasoning, just a different export format.",
        "operationId": "v1_export_case_itp_pdf",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "V1 Export Case Itp Pdf",
        "tags": [
          "v1"
        ]
      }
    },
    "/api/v1/cases/{case_id}/findings": {
      "get": {
        "operationId": "v1_list_case_findings",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "V1 List Case Findings",
        "tags": [
          "v1"
        ]
      },
      "post": {
        "operationId": "v1_add_case_finding",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddFindingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "V1 Add Case Finding",
        "tags": [
          "v1"
        ]
      }
    },
    "/api/v1/cases/{case_id}/findings/{finding_id}": {
      "patch": {
        "operationId": "v1_update_case_finding",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "finding_id",
            "required": true,
            "schema": {
              "title": "Finding Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFindingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "V1 Update Case Finding",
        "tags": [
          "v1"
        ]
      }
    },
    "/api/v1/cases/{case_id}/findings/{finding_id}/rfi": {
      "post": {
        "operationId": "v1_draft_case_rfi",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "finding_id",
            "required": true,
            "schema": {
              "title": "Finding Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "V1 Draft Case Rfi",
        "tags": [
          "v1"
        ]
      }
    },
    "/api/v1/cases/{case_id}/rfis": {
      "get": {
        "operationId": "v1_list_case_rfis",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "V1 List Case Rfis",
        "tags": [
          "v1"
        ]
      }
    },
    "/api/v1/cases/{case_id}/rfis/{rfi_id}": {
      "patch": {
        "operationId": "v1_update_case_rfi",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "rfi_id",
            "required": true,
            "schema": {
              "title": "Rfi Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRfiRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "V1 Update Case Rfi",
        "tags": [
          "v1"
        ]
      }
    },
    "/api/v1/cases/{case_id}/rfis/{rfi_id}/export": {
      "get": {
        "operationId": "v1_export_case_rfi",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "rfi_id",
            "required": true,
            "schema": {
              "title": "Rfi Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "V1 Export Case Rfi",
        "tags": [
          "v1"
        ]
      }
    },
...
  }
}